pyproject.toml
[tool.poetry]
name = "pythonproject5"
version = "0.1.0"
description = ""
authors = ["dinmukhammed <arystambekdimash005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
requests = "^2.31.0"
reportlab = "^4.0.8"
pytelegrambotapi = "^4.14.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pre-commit = "^3.6.0"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''(?x)
  /(
      \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | \.pyc$
'''